Rename geniza.docs to geniza.corpus
