+++ geniza/corpus/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from geniza.docs.models import Library
+from geniza.corpus.models import Library
 
 
 @admin.register(Library)
