+++ geniza/corpus/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class DocsConfig(AppConfig):
-    name = 'docs'
+    name = 'corpus'
